Fix `ArgType.Boolean::convert` function
Since anything which is not exactly "false" was considered to be `true`, it could lead to errors in program behaviour for end users. For example, "0", "no", "hello", and (ridiculously) "False" were parsed as `true`.

This commit changes the function behaviour to case-insensitively parse "true" as `true`, "false" as `false`, and throw ParsingException for anything else.